site: disallow usage of outdoor mode in case outdoor_chanlist is missing
The documentation states the outdoor mode is only enabled in case
outdoor_chanlist is defined in the sites wifi5 configurationsection:

 > If enabled, by setting ``wifi5.outdoor_chanlist``, a number of
 > devices that are commonly installed outdoors will have outdoor
 > mode automatically enabled during their initial setup

This currently does not hold true, as Outdoor mode is always activated
for these devices on first setup, even in case the site does not specify
an outdoor chanlist.

Furthermore, it also does not make sense to show the option neither in
setup-modes basic nor advences configuration view in case the site is
lacking an out door chanlist.

This patch will disable outdoor mode on new installs in case the outdoor
chanlist is missing. It will also not show the respective configuration
options in this case.

Signed-off-by: David Bauer <mail@david-bauer.net>