ipq40xx: Fix automatic ipq-wifi selection for Plasma Cloud devices (#2314)
It was noticed that various devices had not the correct board-2.bin
installed. This was caused by a typo in the package name. The ath10k driver
(unfortunately) is then loading a completely unrelated BDF from the
ath10k-board-qca4019 board-2.bin. It is usually a rather bad idea to use
calibration data from a different board - but the effects depend on the
actual device.

For the PA1200, it was mostly noticed by the bad 2.4GHz performance.

Signed-off-by: Sven Eckelmann <sven@narfation.org>