gluon-core: create interfaces for each "port" in /etc/config/gluon
This commit enables (migration-safe) role assignment for individual "lan"
and "wan" interfaces. This is helpful especially on devices with multiple
"lan" or "wan" interfaces (as they appear for devices with DSA support).

For interfaces in /lib/gluon/lan_ifname, we now create individual
entries in /etc/config/gluon using the "/lan[0]", "/lan[1]", ...
syntax (introduced in the last commit). Before this commit, all interfaces
were referenced in a single interface "/lan" in /etc/config/gluon.

The same is done for wan_ifname and single_ifname.

Migration strategy:

If new interfaces show up during gluon-reconfigure, the assigned roles
of prior existing interfaces in the same category will be applied to the
new interface.