ipq40xx: switch Wave2 firmware to -ct
Use the candelatech firmware for the QCA Wave-2 firmware.

The Qualcomm firmware used for the IPQ401x chip in OpenWrt in 22.03
is experiencing heavily degraded performance due to excessive
retransmits when using A-MSDU. Disabling VHT modes or switching to the
candelatech firmware circumvents this issue.

Apply the same to other Wave-2 platforms in order to keep consistency
with upstream.

Wave-1 chips do not support mesh modes with the -ct firmware, so keep
using the QCA firmware in their case.

Signed-off-by: David Bauer <mail@david-bauer.net>