ath79-generic: (re)add support for ALFA Network AP121F
Gone due to
commit 45c84a1 ("ar71xx: drop target")

Note that it was wrongly marked as device class tiny in
commit 7fd7116e2a86 ("targets: add device-class flags") in the past,
the device has 64MB RAM and not 32MB.

Also, the device has no "led-running" assigned in DTS. The device has
three LEDs: "green:vpn", "green:lan" and "green:wlan". The first LED,
"green:vpn", has a "V" icon and was used to show the VPN connection
status in the vendor firmware. This LED will be used via the newly
added "led-boot" fallback in gluon-setup-mode. But will be unused
during normal operation due to the unassigned "led-running" in DTS.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>