batman-adv: backport routeable mcast address support
With batman-adv 2022.0 support for directed multicast in bridged
setups for routeable multicast addresses was refined - if the
Linux kernel is 5.14 or higher.

This backports the according changes made to the Linux bridge in 5.14
and kernel exports to 5.10, to allow batman-adv to tap into and use
the Linux bridge's Multicast Router Discovery (RFC4286) based multicast
router state. Which enables the batman-adv multicast optimizations not
only for IPv6 link-local multicast addresses (ffX2::/16) but now also
IPv6 routeable multicast addresses (ffXY::/16, with Y > 2) in Gluon.

Note that since batman-adv v2019.3 with a Linux < 5.14 multicast routers
were initially "guessed" by checking for listeners on ff02::2 (all-routers).
Which could potentially overestimate - like in our case:
This approach won't work for us, as every Gluon node listens on ff02::2
on the local-node interface due to Gluon's radvd for local IPv6 prefix
assignments.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>