Dockerfile: don't create gluon user (fixes support for UID != 1000)
The new user `gluon` gets UID 1000 and GID 1000 and cannot write to the
mounted gluon directory unless the calling user has the same UID.

On Linux only the first non admin user gets UID 1000, so docker and podman
fail for all other users.

On macOS there is typically no user with UID 1000, so docker and podman
fail for all users.

When no `gluon` user is created, a default user with the same UID and GID
as the calling user is used and everything works fine.

Signed-off-by: Stefan Weil <sw@weilnetz.de>