generic: disable building openvswitch
openvswitch fails to build with groff v1.23 on the build system.

This groff version hasn't reached a Debian Release yet but is already present
with rolling release distros and causes the gluon build to fail.

Since we don't need openvswitch the best temporary fix seems to disable
building it by default until this get's resolved upstream.