ath79: don't create DIR-825 B1 factory image
Currently the build fails for the D-Link DIR-825 B1. THis is due to the
factory image being size-constrained. The sysupgrade image is not
affected, as OpenWrt now uses a concatenated firmware partition.

To newly install such a device, please use the latest OpenWrt 22.03
factory image and install a Gluon sysupgrade.

Signed-off-by: David Bauer <mail@david-bauer.net>