contrib: Dockerfile: use UID/GID 100 for gluon user
To make the container behave the same regardless of the UIDs used
outside of the container, choose a UID that is unlikely to be used by
the user mounting their build directory the container. 100 should not
conflict with common defaults on Linux (min UID 1000) or MacOS X
(min UID 500). Conveniently, there is already a group "users" at GID
100, so creating a new group can be avoided.