ath79-generic: remove Ubiquiti AirMax targets
Remove due to ticket #2939

These devices risk being bricked in a way that can not be recovered
without running full TFTP or serial recovery, as the flash ends up in a
read-only state with no way of upgrading to a fixed version.

Thus, remove the support for the time being. Don't just flag them as
BROKEN, as this would deliberately brick these boards when installating
such an image.

Signed-off-by: David Bauer <mail@david-bauer.net>