ipq40xx-generic: remove support for AVM FRITZ!Box 7520+7530
remove due to issue https://github.com/freifunk-gluon/gluon/issues/3023

There seem to be issues hindering an updated 7520/7530 to boot properly.

As far as we know the issue only affects openwrt-22.03.

Since the issue results in an unusable device remove support entirely
for now and don't just mark it as broken.