docs: remove unrelated draft link to open-mesh.org in README
This was an idea (thought experiment?) from a Wireless Battlemesh ages
ago between batman-adv devs. Regarding what could be done to interconnect
isolated/separate WiFi mesh clouds on layer 3 without losing best path
metric information, without worsening/compromising on the best route decision.
And as a more sophisticated approach to the batman-adv gateway feature and
a potential replacement for tunneled batman-adv over VPN connections.

Which is the opposite of what the sentence in this Gluon README
is saying. And this never got implemented in either batman-adv nor Gluon.

Therefore removing this link to this old draft page.