scripts: image_customization_lib.lua: fail build without valid customization file
image-customization.lua has replaced GLUON_FEATURES and
GLUON_SITE_PACKAGES, so the file is always required to build a useful
Gluon image. In addition, not only 'file not found' errors were ignored,
but also all other Lua load errors, resulting in a build with empty
image customization when the file was not valid.

Closes #3218