build: extend GLUON_DEBUG with values 0, 1 and 2
Debugging binaries in images that are built with sstrip is extremely
annoying, as not even objdump will work on them.

Switching from sstrip to standard strip increases the ath79 image size
(for some usual Archer C7 v2 configuration I used for testing) by ~70KiB,
or 1.3% of the rootfs size (~1% of the total image size), which is
rarely worth the decreased debuggability.

Extend GLUON_DEBUG to not only distinguish 0 and 1, such that:

- 0 uses sstrip same as before
- 1 switches to regular strip; this is the new default
- 2 includes all debug info, like GLUON_DEBUG=1 did before