gluon-web-network: support handling of active-low PoE passthrough
The Aruba AP-303H offers a switchable PSE output on ETH3. The power
state is controlled using a GPIO pin like seen on CPE210 and Nanostation
alike.

However, the PSE state is active low, which the current code handles as
"Enable PoE Passthrough disable". This is not intuitive, thus support
the active low case in the Gluon Web-UI.

Signed-off-by: David Bauer <mail@david-bauer.net>