scripts: target_config_lib: make opkg a device-specific package
If a package is installed target-wide, but removed for specific devices
(or through image-customization), it is first installed during target
rootfs generation, and then uninstalled again for the derived device
rootfs.

For packages that are marked as essential, removal of the package will
be refused by default, which is usually the correct thing to do to avoid
building broken images.

For opkg, we know that it is not required for regular operation of
Gluon, and removing it may be desirable for devices with small flash.
By never adding opkg to the target list of packages and instead only
installing it as a device-specific package, the uninstall issue can be
avoided.

Removing opkg saves ~38KiB of the squashfs on mips_24kc. Stripping the
opkg database is not supported however, as there is no device-specific
flag for that.