ci: push image for all next branches
Currently the docker image is published for the next branch. If we have
two next branches (e.g. durng a pending OpenWrt release), a branch like
"next-24.10" would not be pushing a docker image.

Match the push action to all branches starting with next. This is
helpful if a downstream CI builds using the Gluon docker image.

Signed-off-by: David Bauer <mail@david-bauer.net>