batman-adv: update to version 2024.3
This updates batman-adv to the latest version, as provided in the
openwrt/routing feed and its master branch. The goal is to prevent the
deprecated 4/32 devices stuck on OpenWrt 19.07 to disable newer
batman-adv features.

The main difference to the openwrt/routing upstream version is:

* noflood mark addition, as used in Gluon (patch 2002)
* always assume absence of multicast routers (patch 9001)
  -> to avoid disabling multicast for packets with a
     routeable multicast destination address
* compat patches to be able to build on OpenWrt 19.07 / Linux 4.14
  (the batman-adv compat hacks approach is too complicated due to
   the compat hacks from wireless-backports, therefore these
   patches apply/revert code inside the net/batman-adv/*.{c,h} directly,
   as a safer way)

This has been tested via https://github.com/T-X/gluon-batman-adv-next/
at Freifunk Lübeck for several months now (with the addition of patches
2001, 2003 and 2004).

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>