libgluonutil: add missing libgen import (#3395)
The import of libgen.h required for basename() was missing. This lead to
undesired behavior on 64 bit systems, where only the upper 32-bit of the
returned pointer was used.

On ARM64 systems such as mediatek-filogic, this lead to a signed extend
of the 32-bit address, leading to a crash of respondd.

Signed-off-by: David Bauer <mail@david-bauer.net>
(cherry picked from commit 3654f83a1da06b9d32d339202f2a9bcf6543d433)