gluon-mesh-batman-adv-brmldproxy: enable proxied querier with mc routers
The whole MLD report proxying only works if there is an MLD querier
somewhere on the mesh side that actually polls for MLD reports. So far
in our tests we configured gateway nodes for the MLD querier job.
To ease the MLD querier setup, to avoid needing to enable an MLD querier
manually on some gateway(s) this makes use of the new adjustments in
brmldproxy that added a bridge interface on top of the proxy dummy
interface:

If batman-adv detected a multicast router behind one of its local client
bridge ports (MRD advertisements are sufficient and recommended) then
enable a potential MLD querier candidate, eligible for MLD querier
election, automatically on this node towards the mesh side.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>