hardware: remove support for D-Link DGS-1210-10P and realtek-rtl838x target
This device should never have been supported as non-broken. It is not a
useful target for batman-adv (as the CPU is too slow for reasonable
throughput; it may be more useful with L3 routing protocols); also, the
autoupdater never worked because of an incorrect image name.

As there it nobody actually running this device (or planning to do so
at the moment), just drop the device instead of fixing the name. The
realtek-rtl838x target becomes empty and is dropped.