gluon-setup-mode: namespace-based implementation for password-less login
Start dropbear in a mount namespace with a readonly overlay mount
replacing passwd and shadow to allow password-less login. This allows us
to drop our custom -f flag.

A login shell wrapper is introduced to move back into the init namespace
for the session, so the overlay mount is invisible to the user and
passwd modifies the right shadow file.