package/gluon-respondd: add target information
For most devices it's quite clear which target they belong to via external information.

For x86 hoewer it's not visible from the outside (besides during upgrades via webserver logs)
to which target a device belongs to.

This is relevant to see for example how many devices are using x86-generic vs. x86-64.

Co-authored-by: Matthias Schiffer <mschiffer@universe-factory.net>