gluon-setup-mode: execute bootcount initscript in setup-mode
Currently Gluon does not execute the bootcount scripts when booting in
setup-mode.

While this is usually not issue after the first installation ahs
completed, it might lead to devices with an A/B partition layout to
switch active partitions after the device has been rebooted multiple
times before the initial configuration.

Fix this by executing the bootcount reset initscript.

Signed-off-by: David Bauer <mail@david-bauer.net>