gluon-core: activate mesh_nolearn
We do not use the multi-hop mesh routing capabilities of 802.11s and
set mesh_fwding=0 in gluon unconditionally.
In turn there is no use in discovering 802.11s mesh paths,
as there are no multi-hop mesh paths due to the lack of forwarding.

While the router identifies this on its own after a few ms,
it is simply not necessary to waste resources on this discovery.

This also works around known bugs in mac80211, which we can avoid
altogether by disabling the learning.

https://github.com/openwrt/openwrt/pull/18657

Related content for further reading:
https://git.sipsolutions.net/iw.git/commit/mesh.c?h=v3.2&id=94ded888ba696a1ec03571d596e96cf944ccc0c8

Co-authored-by: Benjamin Berg <benjamin@sipsolutions.net>
Co-authored-by: Florian Maurer <f.maurer@outlook.de>
Co-authored-by: Jan-Niklas Burfeind <freifunk@aiyionpri.me>