gluon-setup-mode: ignore capacitive buttons on Pulse EX400
The Genexis Pulse EX400 has a capacitive WPS button. This button is
prone to errorous presses, as it is likely to be pressed when picking up
or touching the device.

The device in question has a recessed hardware reset switch for entering
setup-mode. Ignore the capacitive touch button on the device in order to
avoid unwanted activation of setup-mode.

Signed-off-by: David Bauer <mail@david-bauer.net>