mesh-batman-adv: Add parker-vpn to radvd config
When the parker mesh-vpn provider is used `radvd` needs to be configured
in a different way compared to vanilla Gluon.

With this change the different behavior is introduced, if the mesh-vpn
provider is "parker".

In parker every node with an active VPN connection gets an individual
IPv6 prefix (and address). This configuration is placed in `/tmp/addr6`
and `/tmp/prefix6` by the configuration service.

These values are than used to announce the prefix to use and to announce
the node itself as DNS server.