gluon-core: implement detection of client_only_radio
This adds support for the configuration of the third radio on triple radio routers.
The client_only_radio is extended from the existing draft to theoretically also support 2g multi-band.

If there is an additional radio which does not support the mesh channel, we can configure it on a
different channel. In this case, the other radio on the same band supporting the channel turns off its
client radio and is mesh only.