gluon-setup-mode: fix non-interactive dropbear sessions
Our ash wrapper was passing `--login` unconditionally, however this is
incorrect for non-interactive sessions - it does not only result in the
OpenWrt banner to be included in the output of commands like `ssh node
... > output.log`, but also completely breaks scp.

The correct behavior is determined by the value of argv[0] set by
dropbear, which is '-ash' for interactive sessions and 'ash' otherwise,
and interpreted by busybox accordingly. This however poses a problem for
our shell-based wrapper: argv[0] is lost due to the indirection when
passing the script to its interpreter defined by the shebang.

To solve the problem, replace our shell wrapper with a tiny C program,
which preserves the whole argv array when running ash.

Fixes: 18879e28fe74 ("gluon-setup-mode: namespace-based implementation for password-less login")