gluon-client-bridge: configure 6ghz with open SSID instead of OWE ssid
Unencrypted client wifi is not allowed on 6GHz.
The configured SSID to broadcast needs to be the open SSID instead of OWE, as otherwise you end up with two SSID for the same network in the scan list.

Transition networks are hidden identified by the matching relation thru the transition IE. This is not available on 6GHz, thus the broadcasted SSID of the OWE network on 6GHz needs to be the Open SSID on a/g band.