patches: base-files: disable wiphy renaming
The phy renaming is currently only done when the wifi iface is used.
This is too late for gluon - disable it therefore, so that all targets use phyX as well.
This is currently needed for the mediatek-mt7622 target.
Only needed for openwrt-24.10 branch as later releases will be able to handle renaming early enough.